refactor out user

+++ backend-api/src/main/java/com/csc340/localharvest_hub/farm/FarmStatistics.java
@@ -10,25 +10,25 @@ public class FarmStatistics {
     private BigDecimal totalRevenue;
     private BigDecimal monthlyRevenue;
     private Map<String, BigDecimal> revenueByMonth; // Last 6 months
-    
+
     // Subscriber Statistics
     private long totalSubscribers;
     private long activeSubscribers;
     private Map<String, Long> subscribersByMonth; // Subscription growth trend
-    
+
     // Box Statistics
     private int totalBoxes;
     private int activeBoxes;
     private Map<String, Long> mostPopularBoxes; // Box name -> subscriber count
-    
+
     // Rating Statistics
     private double averageOverallRating;
     private double averageFreshnessRating;
     private double averageDeliveryRating;
     private Map<String, Double> ratingsByBox; // Box name -> rating
-    
+
     // Review Statistics
     private long totalReviews;
     private double responseRate; // Percentage of reviews with farmer responses
-    private Map<Integer, Long> ratingDistribution; // Rating -> count
+    private Map<Double, Long> ratingDistribution; // Rating -> count
 }
